backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.2
fastapi==0.135.3
h11==0.16.0
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
watchfiles==1.1.1
yt-dlp==2026.3.17
llvmlite==0.45.1
librosa==0.11.0
numpy<2
scipy==1.15.3
soundfile==0.13.1
replicate==1.0.7
python-dotenv==1.2.2
python-multipart==0.0.20
